seems to be reading the ft... need to make sure it is working, weird thing is that it is turning on the torque to the motors even though i only turned on the power.... i do not know why it is doing that

+++ include/lofaro_types.h
@@ -24,8 +24,8 @@ typedef struct ft_state_def {
         double x;
         double y;
         double voltage;
-        uint8_t raised_x;
-        uint8_t raised_y;
+        int16_t raised_x;
+        int16_t raised_y;
 }__attribute__((packed)) ft_state_def_t;
 
 typedef struct motor_state_def {
